refactor: Change boost variables in BonusCard from float to int
The boost variables in the BonusCard class were changed from float to int to ensure consistent data types. This change improves code readability and maintainability.

fix: Remove commented out code in PlaceCard

The commented out code in the PlaceCard class that applies the effect of the bonus card was removed. This code was no longer needed and was causing confusion.

feat: Add ApplyEffect method in BonusCard

An ApplyEffect method was added to the BonusCard class. This method applies the effect of the bonus card to the appropriate card slots. This improves the functionality of the bonus card and enhances gameplay.

refactor: Update CreateBonusCard method in CardsQuanto

The CreateBonusCard method in the CardsQuanto class was updated to use an int parameter for the initial boost instead of a float. This change aligns the method signature with the updated BoostAmount property in the BonusCard class.

feat: Improve card interaction logic in BonusCard

The card interaction logic in the BonusCard class was improved to only apply the effect to cards of the active player. This ensures that the bonus card only affects the player who played it, enhancing gameplay and strategy.

+++ Assets/src/Decks/CardsQuanto.cs
@@ -188,7 +188,7 @@ public class CardsQuanto : MonoBehaviour
         return climateCard;
     }
 
-    private BonusCard CreateBonusCard(string name, string description, Faction faction, float initialBoost, Effect effect, RowType targetRow, string imageName)
+    private BonusCard CreateBonusCard(string name, string description, Faction faction, int initialBoost, Effect effect, RowType targetRow, string imageName)
     {
         BonusCard bonusCard = ScriptableObject.CreateInstance<BonusCard>();
         bonusCard.Initialize(name, description, TypeofCard.Bonus, faction, initialBoost, effect, targetRow);
